AK-215 [refactor] 주석 추가 및 디렉토리 변경
